backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
alembic
passlib>=1.7.4
bcrypt>=4.0.1
python-jose[cryptography]
sendgrid==6.11.0
python-multipart
httpx>=0.24.1
qrcode[pil]
uvicorn[standard]
apscheduler
numpy 
pandas 
tensorflow-cpu
joblib 
pyarrow
scikit-learn